refactor(web-extension): removes white space from source file name
