+++ packages/web-extension/src/cip30/consumeRemoteAuthenticatorApi.ts
