+++ packages/web-extension/src/cip30/initializeInjectedScript.ts
@@ -1,7 +1,7 @@
 import { Cip30Wallet, WalletProperties, injectGlobal } from '@cardano-sdk/cip30';
 import { Logger } from 'ts-log';
 import { MessengerDependencies, injectedRuntime } from '../messaging';
-import { consumeRemoteAuthenticatorApi } from './consumeRemoteAuthenticatorApi ';
+import { consumeRemoteAuthenticatorApi } from './consumeRemoteAuthenticatorApi';
 import { consumeRemoteWalletApi } from './consumeRemoteWalletApi';
 
 export interface InitializeInjectedDependencies {
